+++ llumnix/backends/vllm_v1/migration_backend.py
@@ -11,7 +11,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import List, Tuple, Optional, Callable, Any
+from typing import List, Tuple, Optional, Callable
 
 import torch
 from func_timeout import func_set_timeout
@@ -32,28 +32,6 @@ from llumnix.constants import (
 from llumnix.utils import random_uuid, ray_get_with_timeout
 
 logger = init_logger(__name__)
-
-
-# Once worker died, proxy actor will not restart.
-@ray.remote(num_cpus=0, max_concurrency=2, max_restarts=-1)
-class ProxyActor:
-    def __init__(self, is_driver_worker: bool, use_ray_spmd_worker: bool):
-        self.is_driver_worker = is_driver_worker
-        self.use_ray_spmd_worker = use_ray_spmd_worker
-
-    def exec_method(self, handle: ray.actor.ActorHandle, *args, **kwargs) -> Any:
-        if self.is_driver_worker and not self.use_ray_spmd_worker:
-            ret = ray_get_with_timeout(
-                handle.execute_engine_method_async.remote(
-                    "execute_driver_worker_method_async", *args, **kwargs
-                )
-            )
-        else:
-            ret = ray_get_with_timeout(
-                handle.execute_method.remote(*args, **kwargs)
-            )
-
-        return ret
 
 
 NUMPY_SUPPORTED_DTYPES = [torch.float32, torch.float16]
